requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0
torchvision>=0.17.0
numpy==1.26.3
Flask==3.0.3
pillow==10.2.0
timm==1.0.11
python-dotenv==1.0.1
pymongo==4.10.1